tests/requirements.txt
pandas==1.0.3
PyYAML==5.3.1
requests==2.23.0
pytest==5.4.1
